pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prescience"
version = "0.1.0"
description = "Camera-based end-of-line counting MVP"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Prescience" }]
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.9",
  "ultralytics>=8.3",
  "torch>=2.2",
  "torchvision>=0.17",
  "typer>=0.12",
  "PyYAML>=6.0",
  "pydantic>=2.6",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "jinja2>=3.1",
  "httpx>=0.27"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
prescience = "prescience.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
